M lib/Session.cpp     specify type cast.
git-svn-id: https://qt-prog.googlecode.com/svn/trunk@28 74b1f884-b478-11dd-b9ef-63de0a379bfc

+++ qtermwidget/lib/Session.cpp
@@ -299,7 +299,7 @@ void Session::run()
   int result = _shellProcess->start(QFile::encodeName(_program),
                                   arguments,
                                   _environment << backgroundColorHint,
-                                  windowId(),
+                                  (ulong)windowId(),
                                   _addToUtmp);
 
   if (result < 0)
